update readme and docs
