+++ cardea/problem_definition/__init__.py
@@ -5,7 +5,7 @@ from cardea.problem_definition.mortality_prediction import MortalityPrediction
 from cardea.problem_definition.predicting_diagnosis import DiagnosisPrediction
 from cardea.problem_definition.prolonged_length_of_stay import ProlongedLengthOfStay
 from cardea.problem_definition.readmission import Readmission
-from cardea.problem_definition.show_noshow_appointment import MissedAppointmentProblemDefinition
+from cardea.problem_definition.show_noshow_appointment import MissedAppointment
 
 __all__ = (
     "ProblemDefinition",
@@ -14,5 +14,5 @@ __all__ = (
     "DiagnosisPrediction",
     "ProlongedLengthOfStay",
     "Readmission",
-    "MissedAppointmentProblemDefinition"
+    "MissedAppointment"
 )
