--- docs/authors.rst
@@ -1,1 +1,0 @@
-.. include:: ../AUTHORS.rst
