+++ docs/basic_concepts/index.rst
@@ -1,3 +1,5 @@
+.. _concepts:
+
 Basic Concepts
 ==============
