+++ docs/getting_started/quickstart.rst
@@ -20,59 +20,60 @@ following command:
 
 .. ipython:: python
 
-    cardea.load_data_entityset()
+    cardea.load_entityset()
+    cardea.es
 
 You can see the list of problem definitions and select one with the following commands:
 
 .. ipython:: python
 
     cardea.list_problems()
-    problem = cardea.select_problem('MissedAppointmentProblemDefinition')
 
-Then, you can perform the AutoML steps and take advantage of Cardea:
+From there, you can select the prediction problem you aim to solve by specifying the name of the class, which in return gives us the ``label_times`` of the problem. 
 
-1. Extracting features (automated feature engineering), using the following commands:
+.. ipython:: python
 
-   .. ipython:: python
+    label_times = cardea.select_problem('MissedAppointment')
+    label_times.head()
 
-       feature_matrix = cardea.generate_features(problem[:1000])  # a subset
-       feature_matrix = feature_matrix.sample(frac=1)  # shuffle
-       y = list(feature_matrix.pop('label'))
-       X = feature_matrix.values
+Then, you can perform the AutoML steps and take advantage of Cardea.
 
-2. Modeling, optimizing hyperparameters and finding the most optimal model using the following commands:
+Cardea extracts features through automated feature engineering by supplying the ``label_times`` pertaining to the problem you aim to solve, using the following commands:
 
-   .. ipython:: python
+ .. ipython:: python
+     :okwarning:
 
-       pipeline = [
-           ['sklearn.ensemble.RandomForestClassifier'],
-           ['sklearn.naive_bayes.MultinomialNB'],
-           ['sklearn.neighbors.KNeighborsClassifier']
-       ]
+     feature_matrix = cardea.generate_features(label_times[:1000])  # a subset
+     feature_matrix.head()
 
-       result = cardea.execute_model(
-           feature_matrix=X,
-           target=y,
-           primitives=pipeline
-       )
+Once we have the features, we can now split the data into training and testing
+
+ .. ipython:: python
+     :okwarning:
+
+     y = list(feature_matrix.pop('label'))
+     X = feature_matrix.values
+
+     X_train, X_test, y_train, y_test = cardea.train_test_split(
+         X, y, test_size=0.2, shuffle=True)
+
+
+Now that we have our feature matrix properly divided, we can use to train our machine learning pipeline, Modeling, optimizing hyperparameters and finding the most optimal model is done using the following commands:
+
+ .. ipython:: python
+     :okwarning:
+
+     cardea.select_pipeline('Random Forest')
+     cardea.fit(X_train, y_train)
+     y_pred = cardea.predict(X_test)
 
 
 Finally, you can see accuracy results using the following commands:
 
-.. ipython:: python
+ .. ipython:: python
+     :okwarning:
+     
+     cardea.evaluate(X, y, test_size=0.2, metrics=['Accuracy', 'F1 Macro'])
 
-    import pandas as pd
-    from sklearn.metrics import accuracy_score
-
-    y_test = []
-    y_pred = []
-
-    for i in range(0, 10):
-        y_test.extend(result['pipeline0']['folds'][str(i)]['Actual'])
-        y_pred.extend(result['pipeline0']['folds'][str(i)]['predicted'])
-
-    y_test = pd.Categorical(pd.Series(y_test)).codes
-    y_pred = pd.Categorical(pd.Series(y_pred)).codes
-    accuracy_score(y_test, y_pred)
 
 .. _Medical Appointment No Shows: https://www.kaggle.com/joniarroba/noshowappointments
