+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
-envlist = py35, py36, lint, docs
+envlist = py37, py36, lint, docs
 
 
 [travis]
 python =
     3.6: py36, lint, docs
-    3.5: py35
+    3.7: py37
 
 
 [testenv]
